+++ old_src/solver/basic_fish_solver.h
@@ -9,7 +9,7 @@
 #include "utils.h"
 
 typedef struct {
-  uint16_t indicies;
+  uint8_t indicies;
   int baseIndex;
 } FishIndicies;
 
@@ -25,9 +25,9 @@ StepNode* findBasicJellyfish(Puzzle* puzzle, StepNode* head);
 
 StepNode* findBasicFishByRow(Puzzle* puzzle, int fishSize, Strategy stratUsed, StepNode* head);
 StepNode* findBasicFishByCol(Puzzle* puzzle, int fishSize, Strategy stratUsed, StepNode* head);
-void findCandidateIndiciesInHouse(uint16_t* house, uint16_t candidateMask, uint16_t* indicies, int indiciesSize);
-int countCandidateInHouse(uint16_t* house, uint16_t candidateMask, int maxCount);
-StepNode* removeCandidatesFromColExcept(Puzzle* puzzle, uint16_t candidateMask, int colIndex, int* exceptIndicies, int exceptSize, Strategy stratUsed, StepNode* head);
-StepNode* removeCandidatesFromRowExcept(Puzzle* puzzle, uint16_t candidateMask, int rowIndex, int* exceptIndicies, int exceptSize, Strategy stratUsed, StepNode* head);
+void findCandidateIndiciesInHouse(uint8_t* house, uint8_t candidateMask, uint8_t* indicies, int indiciesSize);
+int countCandidateInHouse(uint8_t* house, uint8_t candidateMask, int maxCount);
+StepNode* removeCandidatesFromColExcept(Puzzle* puzzle, uint8_t candidateMask, int colIndex, int* exceptIndicies, int exceptSize, Strategy stratUsed, StepNode* head);
+StepNode* removeCandidatesFromRowExcept(Puzzle* puzzle, uint8_t candidateMask, int rowIndex, int* exceptIndicies, int exceptSize, Strategy stratUsed, StepNode* head);
 
 #endif // SRC_SOLVER_BASIC_FISH_SOLVER_H
