+++ old_src/solver/finned_fish_solver.h
@@ -13,7 +13,7 @@ typedef struct {
   int housesWithCandidateCount;
   FishBases* allFishBases;
   FishBases* potentialBaseCombo;
-  uint16_t* allCandidates;
+  uint8_t* allCandidates;
   int candidate;
   bool byRow;
   int* indiciesToRemove;
@@ -27,8 +27,8 @@ StepNode* findFinnedJellyfish(Puzzle* puzzle, StepNode* head);
 StepNode* findFinnedFishByRow(Puzzle* puzzle, int fishSize, Strategy stratUsed, StepNode* head);
 StepNode* findFinnedFishByCol(Puzzle* puzzle, int fishSize, Strategy stratUsed, StepNode* head);
 bool searchFishRecursively(FinnedFishContext* context, int startIndex, int depth, int fishSize);
-uint16_t findRowsWithCandidate(uint16_t* puzzleCandidates, int candidate);
-uint16_t findColsWithCandidate(uint16_t* puzzleCandidates, int candidate);
+uint8_t findRowsWithCandidate(uint8_t* puzzleCandidates, int candidate);
+uint8_t findColsWithCandidate(uint8_t* puzzleCandidates, int candidate);
 void getBlockIndicies(int blockX, int blockY, int* blockIndices);
 
 #endif // SRC_SOLVER_FINNED_FISH_SOLVER_H
